move search state down to Search.js
