+++ train-tracker/src/Search.js
@@ -1,14 +1,36 @@
 import './styles/Search.css';
+import {useState} from 'react'
 
-function Search({
-    number = "", setNumber = function(){},
-    route = "", setRoute = function(){}, routes = <option value={""}>{}</option>,
-    station = "", setStation = function(){}, stations = <option value={""}>{}</option>,
-    upcoming = false, setUpcoming = function(){},
-    searchFun = function(){}, clearFun = function (){},
-    fromStation = "", setFromStation = function(){},
-    toStation = "", setToStation = function(){}
-}){
+function Search({searchFun, routes, stations}){
+    const [selectedNumber, setSelectedNumber] = useState("");
+    const [selectedRoute, setSelectedRoute] = useState("");
+    const [selectedStation, setSelectedStation] = useState("");
+    const [upcoming, setUpcoming] = useState(false);
+    const [fromStation, setFromStation] = useState("");
+    const [toStation, setToStation] = useState("");
+
+    function handleNumber(e){ setSelectedNumber(e.target.value); }
+    function handleRoute(e){ setSelectedRoute(e.target.value); }
+    function handleStation(e){ setSelectedStation(e.target.value); }
+    function handleUpcoming(e){ setUpcoming(e.target.checked); }
+    function handleFromStation(e){ setFromStation(e.target.value); }
+    function handleToStation(e){ setToStation(e.target.value); }
+
+    const search = () =>{
+        searchFun(selectedNumber, selectedRoute, selectedStation, upcoming, fromStation, toStation);
+    }
+
+    const clearSearch = () => {
+        setSelectedNumber("");
+        setSelectedRoute("");
+        setSelectedStation("");
+        setUpcoming(false);
+        setFromStation("");
+        setToStation("");
+
+        searchFun(selectedNumber, selectedRoute, selectedStation, upcoming, fromStation, toStation);
+    }
+
     return (
         <form className='form'>
                 <div className='top-label'>
@@ -16,34 +38,33 @@ function Search({
                 </div>
                 <span className="select-label">
                         Train Number:
-                        <input value={number} onChange={setNumber} type="number" placeholder="Enter Train Number"></input>
+                        <input value={selectedNumber} onChange={handleNumber} type="number" placeholder="Enter Train Number"></input>
                     </span>
                 <span className="select-label">
                         Route:
-                        <select value={route} onChange={setRoute} children={routes}></select>
+                        <select value={selectedRoute} onChange={handleRoute} children={routes}></select>
                     </span>
                 <span className="select-label">By station: </span>
-                    <select value={station} onChange={setStation} children={stations}></select>
+                    <select value={selectedStation} onChange={handleStation} children={stations}></select>
                     <span className="select-label">
                         Upcoming trains only: 
-                        <input checked={upcoming} onChange={setUpcoming} type="checkbox" ></input>
+                        <input checked={upcoming} onChange={handleUpcoming} type="checkbox" ></input>
                     </span>
                 
                 <span>
-                    <button type="button" onClick={searchFun}>Search</button>
-                    <button type="button" onClick={clearFun}>Clear</button>
+                    <button type="button" onClick={search}>Search</button>
+                    <button type="button" onClick={clearSearch}>Clear</button>
                 </span>
                 
-
                 <label className="optional-criteria-label">
                     Optional criteria:
                     <span className="select-label">
                         From:
-                        <select value={fromStation} onChange={setFromStation} children={stations}></select>
+                        <select value={fromStation} onChange={handleFromStation} children={stations}></select>
                     </span>
                     <span className="select-label">
                         To:
-                        <select value={toStation} onChange={setToStation} children={stations}></select>
+                        <select value={toStation} onChange={handleToStation} children={stations}></select>
                     </span>
                 </label>
               </form>
